rename and clear
