+++ src/pages/SettingsPage.tsx
@@ -1,13 +1,11 @@
 import React, { FC } from 'react';
 import { WithHeadbarLayout } from '@/layouts/WithHeadbarLayout';
 import { GameSettings } from '@/modules/GameSettings';
-import { Field } from '@/modules/Field';
 
 export const SettingsPage: FC = () => {
   return (
     <WithHeadbarLayout>
       <GameSettings />
-      {/* <Field presentation /> */}
     </WithHeadbarLayout>
   );
 };
